chore(refactor) moves resetpassword test in user route test structure [Finishes #167730744]
